changed conversions to float 32

+++ basic_rythm_model/single_class/convert_mlpackage_single.py
@@ -1,22 +1,22 @@
-import torch  # lub importuj klasę tak samo jak dla multi
+import torch
 import numpy as np
 import coremltools as ct
-from main_single import N_CLASSES , SEG_LEN , AGE_MEAN , SAVE_NAME
+from main_single import N_CLASSES, SEG_LEN, AGE_MEAN, SAVE_NAME
 from main_single import SE_MobileNet1D_noLSTM
 
 # === Wczytanie modelu ===
 model = SE_MobileNet1D_noLSTM(num_classes=N_CLASSES)
-model.load_state_dict(torch.load(SAVE_NAME + ".pt" , map_location="cpu"))
+model.load_state_dict(torch.load(SAVE_NAME + ".pt", map_location="cpu"))
 model.eval()
 
 # === Przykładowe dane wejściowe ===
-example_input = torch.randn(1, 1, SEG_LEN)
-example_demo = torch.tensor([[AGE_MEAN, 1.0]])  # np. wiek i płeć
+example_input = torch.randn(1, 1, SEG_LEN, dtype=torch.float32)  # ✅ wymuszenie float32
+example_demo = torch.tensor([[AGE_MEAN, 1.0]], dtype=torch.float32)  # ✅ wymuszenie float32
 
 # === Trace modelu ===
 traced = torch.jit.trace(model, (example_input, example_demo))
 
-# === Konwersja do Core ML ===
+# === Konwersja do Core ML (ML Program, float32) ===
 mlmodel = ct.convert(
     traced,
     inputs=[
@@ -26,5 +26,5 @@ mlmodel = ct.convert(
     convert_to="mlprogram"
 )
 
-# === Zapis ===
+# === Zapis modelu ===
 mlmodel.save(SAVE_NAME + ".mlpackage")
